+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/util/extensions/UiExtensions.kt
@@ -2,6 +2,7 @@ package net.nymtech.nymvpn.util.extensions
 
 import android.annotation.SuppressLint
 import android.content.Context
+import android.text.format.DateUtils
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.drawWithContent
@@ -18,23 +19,23 @@ import androidx.navigation.NavDestination.Companion.hasRoute
 import androidx.navigation.NavDestination.Companion.hierarchy
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import net.nymtech.nymvpn.NymVpn
-import net.nymtech.nymvpn.ui.Route
-import kotlin.reflect.KClass
 import net.nymtech.nymvpn.R
+import net.nymtech.nymvpn.ui.Route
 import net.nymtech.nymvpn.ui.theme.CustomColors
 import net.nymtech.vpn.backend.Tunnel
 import net.nymtech.vpn.model.NymGateway
+import nym_vpn_lib.VpnException
 import nym_vpn_lib_types.EntryPoint
+import nym_vpn_lib_types.ErrorStateReason
 import nym_vpn_lib_types.ExitPoint
 import nym_vpn_lib_types.GatewayType
 import nym_vpn_lib_types.Score
-import nym_vpn_lib_types.ErrorStateReason
-import nym_vpn_lib.VpnException
 import timber.log.Timber
 import java.time.Instant
 import java.time.ZoneId
 import java.time.format.DateTimeFormatter
-import java.util.*
+import java.util.Locale
+import kotlin.reflect.KClass
 
 fun Dp.scaledHeight(previewScale: Float = 1f): Dp = if (NymVpn.isInitialized) {
 	NymVpn.instance.resizeHeight(this)
@@ -162,18 +163,22 @@ fun VpnException.toUserMessage(context: Context): String {
 fun List<NymGateway>.scoreSorted(mode: Tunnel.Mode): List<NymGateway> {
 	return this.sortedBy {
 		when (mode) {
-			Tunnel.Mode.FIVE_HOP_MIXNET -> it.mixnetScore
-			Tunnel.Mode.TWO_HOP_MIXNET -> it.wgScore
+			Tunnel.Mode.FIVE_HOP_MIXNET -> it.mixnetScore ?: Score.OFFLINE
+			Tunnel.Mode.TWO_HOP_MIXNET -> it.wgScore ?: Score.OFFLINE
 		}
 	}
 }
 
+fun toDisplayCountry(twoLetterIsoCountryCode: String): String {
+	return Locale(twoLetterIsoCountryCode, twoLetterIsoCountryCode).displayCountry
+}
+
 fun EntryPoint.Country.toDisplayCountry(): String {
-	return Locale(this.twoLetterIsoCountryCode, this.twoLetterIsoCountryCode).displayCountry
+	return toDisplayCountry(twoLetterIsoCountryCode)
 }
 
 fun ExitPoint.Country.toDisplayCountry(): String {
-	return Locale(this.twoLetterIsoCountryCode, this.twoLetterIsoCountryCode).displayCountry
+	return toDisplayCountry(twoLetterIsoCountryCode)
 }
 
 @Composable
@@ -231,6 +236,17 @@ fun formatUtcString(utcString: String?): String {
 			formatter.format(instant)
 		} ?: "--"
 	} catch (e: Exception) {
+		"--"
+	}
+}
+
+fun relativeTimeSpan(utcString: String?): String {
+	return try {
+		utcString?.let {
+			val instant = Instant.parse(it)
+			DateUtils.getRelativeTimeSpanString(instant.toEpochMilli()).toString()
+		} ?: "--"
+	} catch (_: Exception) {
 		"--"
 	}
 }